Implement skeleton for new Add command
This commit implements a skeleton for the addition of various fields to
the Add command in preparation for v1.2, as indicated in
`docs/UserGuide.md`.

The following changes were made:
1. `address` has been renamed to `homeAddress`
2. The following fields were added: `workAddress`, `quarantineAddress`,
   `shnPeriod`, `caseNumber`, `nextOfKinName`, `nextOfKinPhone`,
   `nextOfKinAddress`
3. `tags` was deprecated and will be removed when the changes are
   integrated.
4. Related variables and classes have been modified accordingly. All
   new fields are of type `Object` and set to `null` by default.
   Currently, equality checks exclude these fields.

A list of tasks to be carried out during integration has been added into
the codebase at relevant points, marked by `TODO` comments.